pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-chunk-visualizer"
version = "0.1.0"
description = "Interactive local-first RAG visualization and debugging app"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "streamlit",
    "plotly",
    "pandas",
    "numpy",
    "sentence-transformers",
    "lancedb",
    "scikit-learn",
    "umap-learn",
    "python-dotenv",
    "ollama",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"